Move common code to module_utils
